:art: remove ` imported` and modified version check
I need to remove the ` imported` because it is doing absolutely nothing. And the other thing was on LN 49 the text isn't explanatory. The if condition is also broken. It is currently applying  `-gt` greater than. I replace this in this commit with `-ge` which stands for greater or equal. See the reference here: https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_comparison_operators?view=powershell-7.2#long-description